Add files via upload
Instructions for an AVR GNU cross-compilation environment, with a focus on upgrading AVRDUDE and the low-cost Olimex AVR-ISP-MK2 programmer. Focuses on OS X + MacPorts environment, but may be useful to anyone who wants a clean cross-compilation environment on Linux / BSD / etc. The toolchain is based on the latest stable software releases as of 20170518. e.g., gcc-7.1.0. This is a guide for those comfortable with command line and shell-based hacking.